Struct ayaka_model::GameViewModel

pub struct GameViewModel<S: SettingsManager, M: RawModule + Send + Sync + 'static> { /* private fields */ }

A view model of Ayaka. It manages all settings and provides high-level APIs.

Implementations

impl<S: SettingsManager, M: RawModule + Send + Sync + 'static> GameViewModel<S, M>

pub fn new(settings_manager: S) -> Self

Create a GameViewModel with a settings manager.

pub fn open_game<'a>( &'a mut self, context: Context<M> ) -> impl Future<Output = Result<()>> + Stream<Item = OpenGameStatus> + 'a

Open the game with context.

pub fn context(&self) -> &Context<M>

The [Context], should be called after Self::open_game.

pub fn context_mut(&mut self) -> &mut Context<M>

The [Context], should be called after Self::open_game.

pub fn record(&self) -> &ActionRecord

The current ActionRecord.

pub fn settings(&self) -> &Settings

The loaded Settings.

pub fn set_settings(&mut self, settings: Settings)

Set the Settings.

pub fn records(&self) -> &[ActionRecord]

The loaded ActionRecords.

pub fn global_record(&self) -> &GlobalRecord

The loaded GlobalRecord.

pub fn global_record_mut(&mut self) -> &mut GlobalRecord

The loaded GlobalRecord.

pub fn avaliable_locale(&self) -> impl Iterator<Item = &Locale>

Get the avaliable locales from paragraphs.

pub fn init_new(&mut self)

Start a new game. The state of the model after this call is actually invalid. You should call next_run immediately after this call, to ensure there’s content in the game, and switch to the first line of the first paragraph.

pub fn init_context(&mut self, record: ActionRecord)

Start a game with record.

pub fn init_context_by_index(&mut self, index: usize)

Start a game with the index of records.

pub fn next_run(&mut self) -> bool

Step to the next run.

pub fn next_back_run(&mut self) -> bool

Step back to the last run.

pub fn current_run(&self) -> Option<&RawContext>

Get the current [RawContext].

pub fn current_title(&self) -> Option<&String>

Get the current paragraph title.

pub fn current_action(&self) -> Option<Action>

Get the current action by language.

pub fn current_actions(&self) -> Option<(Action, Option<Action>)>

Get the current action by language and secondary language.

pub fn switch(&mut self, i: usize)

Choose a switch item by index.

pub fn save_current_to(&mut self, index: usize)

Save current ActionRecord to the records.

pub fn save_settings(&self) -> Result<()>

Save all settings and records.

pub fn current_visited(&self) -> bool

Determine if current run has been visited.

pub fn records_text(&self) -> impl Iterator<Item = ActionText> + '_

Get the last action text from each record.

pub fn current_history( &self ) -> impl DoubleEndedIterator<Item = (Action, Option<Action>)> + '_

Get the current history by language and secondary language.